perf(renderer): store reflector handles once
